backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
motor==3.6.0
pydantic==2.9.2
requests==2.32.3
python-multipart==0.0.12
websockets==13.0.1
pymongo>=4.6.0,<4.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.9.0.post0
beautifulsoup4==4.12.3
feedparser==6.0.12
pytz==2024.2
geopy==2.4.1
pandas>=2.2.0,<3
numpy>=2.1.0,<3
#python-otx>=1.5.0  # Temporarily commented out as not available
#virustotal-api>=1.1.11  # May not be available on all platforms
#abuseipdb-python>=0.1.2  # May not be available on all platforms
python-slugify==8.0.4
